Move preSig from urbit/ -> lib/urbit.ts
