+++ packages/api/src/client/branch.ts
@@ -2,8 +2,8 @@ import { valid } from '@urbit/aura';
 
 import { getPostInfoFromWer } from '../client/harkApi';
 import { createDevLogger } from '../lib/logger';
+import { preSig } from '../lib/urbit';
 import { getConstants } from '../types/constants';
-import { preSig } from '../urbit';
 import { normalizeUrbitColor } from './utils';
 
 const logger = createDevLogger('branch', false);
