+++ packages/app/contexts/ship.tsx
@@ -1,5 +1,5 @@
 import crashlytics from '@react-native-firebase/crashlytics';
-import { preSig } from '@tloncorp/api/urbit';
+import { preSig } from '@tloncorp/api/lib/urbit';
 import { AnalyticsEvent, createDevLogger } from '@tloncorp/shared';
 import { ShipInfo, storage } from '@tloncorp/shared/db';
 import type { ReactNode } from 'react';
