+++ packages/app/hooks/useCurrentUser.native.ts
@@ -1,14 +1,10 @@
-import * as urbit from '@tloncorp/api/urbit';
+import { preSig } from '@tloncorp/api/lib/urbit';
 
 import { useShip } from '../contexts/ship';
 
 export function useCurrentUserId() {
   const { ship, contactId } = useShip();
-  const userId = contactId
-    ? urbit.preSig(contactId)
-    : ship
-      ? urbit.preSig(ship)
-      : null;
+  const userId = contactId ? preSig(contactId) : ship ? preSig(ship) : null;
   // this should be redundant (contactId should always be present), but throw
   // error to avoid ever having a missing current user when authenticated
   if (!userId) {
